更新了cate

+++ src/Https/Controllers/CommodityCateController.php
@@ -4,9 +4,12 @@ namespace SimpleShop\Cate\Https\Controllers;
 
 use App\Http\Controllers\Controller;
 use SimpleShop\Cate\Contracts\Cate;
+use SimpleShop\Cate\Exceptions\CommodityCateException;
+use SimpleShop\Cate\Exceptions\ExceptionCode;
 use SimpleShop\Cate\Https\Requests\Api\ListRequest;
 use SimpleShop\Cate\Https\Requests\Api\SubmitRequest;
 use SimpleShop\Cate\Https\Traits\ReturnFormat;
+use SimpleShop\Commodity\Models\ShopGoodsModel;
 
 class CommodityCateController extends Controller
 {
@@ -51,6 +54,7 @@ class CommodityCateController extends Controller
      *
      *
      * @param SubmitRequest $request
+     *
      * @return \Illuminate\Http\JsonResponse
      */
     public function store(SubmitRequest $request)
@@ -78,6 +82,7 @@ class CommodityCateController extends Controller
      * Show the form for editing the specified resource.
      *
      * @param  int $id
+     *
      * @return \Illuminate\Http\Response
      */
     public function edit($id)
@@ -94,9 +99,10 @@ class CommodityCateController extends Controller
      *
      * @return \Illuminate\Http\JsonResponse
      */
-    public function update(SubmitRequest $request, $id)
+    public function update(SubmitRequest $request, $id = null)
     {
-        $this->service->update($id, $request->all());
+        $id = is_null($id) ?? $request->input('id');
+        $this->service->update($id, $request->except(['userType']));
 
         return $this->success();
     }
@@ -108,8 +114,19 @@ class CommodityCateController extends Controller
      *
      * @return \Illuminate\Http\JsonResponse
      */
-    public function destroy($id)
+    public function destroy($id, $version = null)
     {
+        $result = ShopGoodsModel::where('cate_id', $id)->first(['id']);
+        if (! is_null($result)) {
+            throw new CommodityCateException("该分类已经被商品使用", ExceptionCode::DATA_USED);
+        }
+
+        $result = $this->service->getChildren($id);
+
+        if ($result->isNotEmpty()) {
+            throw new CommodityCateException("该分类已经被其他分类使用", ExceptionCode::DATA_USED);
+        }
+        $this->service = clone $this->service;
         $this->service->destroy($id);
 
         return $this->success();
